requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==20.0.4
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
fonttools==4.47.2
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
neo4j==5.17.0
numpy==1.24.4
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
setuptools==58.1.0
six==1.16.0
tzdata==2023.4
Werkzeug==3.0.1
requests==2.20.0
